MeterSuspension

+++ Source/Fasetto.Word.Core/ApiModels/Services/BulkReconDetailResultApiModel.cs
@@ -33,7 +33,7 @@ namespace Fasetto.Word.Core
         /// <summary>
         ///Consumption for meter over defined period
         /// </summary>
-        public float Volume { get; set; }
+        public decimal Volume { get; set; }
 
         /// <summary>
         ///Meter reading (calculated) for start of measurement period
